Prevent StackOverflowError
StackOverflowError at handle() is seen when using this code to tail a file and wait for some time. It seems tail recursion optimization is not applied by the Scalar compiler because handle() is not marked final. I've modified this part and added @tailrec to further made sure the optimization is applied as expected. So far I do not see any StackOverflowError.